requirements.txt
amqp==5.3.1
app==0.0.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2.13
django-timezone-field==7.1
djangorestframework==3.16.0
entrypoints==0.4
exceptiongroup==1.3.0
idna==3.10
iniconfig==2.1.0
kombu==5.5.2
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-django==4.11.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
